+++ class_settings/options.py
@@ -2,7 +2,11 @@ from django.conf import global_settings
 
 
 class Options:
-    defaults = {"default_settings": global_settings, "env_prefix": "django"}
+    defaults = {
+        "default_settings": global_settings,
+        "inject_settings": False,
+        "env_prefix": "django",
+    }
 
     def __init__(self, meta):
         for option, default in self.defaults.items():